pyproject.toml
[project]
name = "med-center"
version = "0.1.0"
description = ""
authors = [
    {name = "Floccinau",email = "argonavt28@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-ninja (>=1.4.1,<2.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "celery[redis] (>=5.5.2,<6.0.0)",
    "django-celery-beat (>=2.8.0,<3.0.0)",
    "redis (>=5.0.1,<6.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
pre-commit = "^4.2.0"

[tool.isort]
# Global
multi_line_output = 3
include_trailing_comma = true
line_length = 79
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2

# Local
src_paths = ["core"]

skip_glob = [
    ".git",
    ".idea",
    ".vscode",
    ".venv",
    "venv"
]

# Known
known_django = "django"
known_ninja = "ninja"
known_first_party = ["core"]

# Sections (imports order)
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "NINJA",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER"
]